+++ web/tests/lib/events.cjs
@@ -894,6 +894,7 @@ exports.fixtures = {
             creator_id: null,
             date_created: fake_now,
             description: "mobile folks",
+            rendered_description: "<p>mobile folks</p>",
             members: [1],
             is_system_group: false,
             direct_subgroup_ids: [2],
@@ -942,6 +943,7 @@ exports.fixtures = {
         data: {
             name: "Frontend",
             description: "All Frontend people",
+            rendered_description: "<p>All Frontend people</p>",
             can_manage_group: 2,
         },
     },