+++ zerver/lib/import_realm.py
@@ -813,6 +813,7 @@ def bulk_import_named_user_groups(data: TableData) -> None:
             group["realm_for_sharding_id"],
             group["name"],
             group["description"],
+            group["rendered_description"],
             group["is_system_group"],
             group["can_add_members_group_id"],
             group["can_join_group_id"],
@@ -828,7 +829,7 @@ def bulk_import_named_user_groups(data: TableData) -> None:
 
     query = SQL(
         """
-        INSERT INTO zerver_namedusergroup (usergroup_ptr_id, realm_id, name, description, is_system_group, can_add_members_group_id, can_join_group_id,  can_leave_group_id, can_manage_group_id, can_mention_group_id, can_remove_members_group_id, deactivated, date_created)
+        INSERT INTO zerver_namedusergroup (usergroup_ptr_id, realm_id, name, description, rendered_description, is_system_group, can_add_members_group_id, can_join_group_id,  can_leave_group_id, can_manage_group_id, can_mention_group_id, can_remove_members_group_id, deactivated, date_created)
         VALUES %s
         """
     )