+++ zerver/views/user_groups.py
@@ -74,6 +74,9 @@ def add_user_group(
     group_settings_map = {}
     request_settings_dict = locals()
     nobody_group = get_system_user_group_by_name(SystemGroups.NOBODY, user_profile.realm_id)
+    if "\n" in description:
+        # We don't allow newline characters in group descriptions.
+        description = description.replace("\n", " ")
     for setting_name, permission_config in NamedUserGroup.GROUP_PERMISSION_SETTINGS.items():
         if setting_name not in request_settings_dict:  # nocoverage
             continue
@@ -161,6 +164,9 @@ def edit_user_group(
         do_update_user_group_name(user_group, name, acting_user=user_profile)
 
     if description is not None and description != user_group.description:
+        if "\n" in description:
+            # We don't allow newline characters in group descriptions.
+            description = description.replace("\n", " ")
         do_update_user_group_description(user_group, description, acting_user=user_profile)
 
     request_settings_dict = locals()